+++ video-element-srcset.js
@@ -1,7 +1,6 @@
+// requestAnimationFrame polyfill by Erik Möller. fixes from Paul Irish and Tino Zijdel
 // http://paulirish.com/2011/requestanimationframe-for-smart-animating/
 // http://my.opera.com/emoller/blog/2011/12/20/requestanimationframe-for-smart-er-animating
-
-// requestAnimationFrame polyfill by Erik Möller. fixes from Paul Irish and Tino Zijdel
 
 // MIT license
 
@@ -42,7 +41,7 @@
  * @param   {array}     videoSources    An array of arrays containing a media string, a video src (URL) string
  *                                      and and string to signaling if the src was fully loaded or not.
  *                                      The key is the media string.
- *          @param   {string}     videoSources[i][0]    a media string to be tested against
+ *          @param   {string}     videoSources[i][0]    device-based breakpoint names to be tested against
  *          @param   {string}     videoSources[i][1]    a video src, it must be an URL
  *          @param   {string}     videoSources[i][2]    a media string to be signal if the src was loaded fully
  *                                                      Values: 'loaded' or 'not-loaded'.
@@ -50,11 +49,11 @@
  * @return {void}
  */
 
-// TODO use of service work and storageCache to avoid unnecessary loading of video files
-var videoElementSrcset = function (videoElement, videoSources) {
+ var videoElementSrcset = function (videoElement, videoSources) {
 
     console.info('videoElementSrcset init');
 
+    // declare our variables
     var breakpoint, previousBreakpoint, activeVideoBreakpoint, timeout, videoSourceElement, activeSource, videoSrcSet = '';
 
     // if videoElement is not provided, get the first video element in the document
@@ -70,6 +69,7 @@ var videoElementSrcset = function (videoElement, videoSources) {
     // Public APIs
     var publicAPIs = {};
 
+    // breakpoint weights, used to avoid unnecessaries video loading
     var breakpointWeight = {
         '"xsmall"': 1,
         '"small"':  2,
@@ -80,6 +80,7 @@ var videoElementSrcset = function (videoElement, videoSources) {
 
     console.log('breakpointWeight initial', breakpointWeight);
 
+    // install Service Worker to better cache management
     var installWorker = function () {
         console.log('init installWorker');
         navigator.serviceWorker.register('./sw-video-cache.js', { scope: './' }).then(function (registration) {
@@ -111,6 +112,8 @@ var videoElementSrcset = function (videoElement, videoSources) {
     // if there is no sources, bail and logs to console
     if (!videoSources || 0 === videoSources.length) {
         console.info('videoSources paramenter is empty');
+
+
         videoSourceElement = document.querySelector('video source');
         if (0 === videoSourceElement.length) {
             console.info('there is no source element to this video element');
@@ -119,7 +122,7 @@ var videoElementSrcset = function (videoElement, videoSources) {
             console.info('there is A source element to this video element', videoSourceElement);
         }
 
-
+        // get our information from markup
         videoSrcSet = videoSourceElement.getAttribute('data-srcset');
         console.info('videoSrcSet getattribute = ', videoSrcSet );
 
@@ -144,7 +147,6 @@ var videoElementSrcset = function (videoElement, videoSources) {
 
             console.info('index ' + i + 'type of  components[0] = ', typeof components[0]);
             videoSources['"' + components[0] + '"'] = [components[1], 'not-loaded'] ;
-            //videoSources.push([components[0], [ [components[1], 'not-loaded'] ] ] );
 
             console.info('videoSources index ' + i + '=', videoSources['"' + components[0] + '"'] );
         }
@@ -155,7 +157,7 @@ var videoElementSrcset = function (videoElement, videoSources) {
         console.info('videoSources is NOT empty and is =', videoSources);
     }
 
-
+    // window.resize listener
     var breakpointListener = function () {
         console.info('breakpointListener init');
         console.log('breakpoint no breakpointListener init = ', breakpoint);
@@ -174,60 +176,75 @@ var videoElementSrcset = function (videoElement, videoSources) {
         return;
     }
 
-    publicAPIs.evaluateBreakpoing = function(){
-
-    }
-
+    // manager the source swap
     publicAPIs.manageSrc = function ( breakpoint ) {
         console.log('publicAPIs.manageSrc init');
         console.log('publicAPIs.manageSrc init breakpoint = ', breakpoint);
         console.log('publicAPIs.manageSrc init previousBreakpoint = ', previousBreakpoint);
+
+        // what is the active src?
         activeSource = videoSourceElement.getAttribute('src');
         console.log('publicAPIs.manageSrc activeSource = ', activeSource);
         console.log('publicAPIs.manageSrc videoSources= ', videoSources);
         console.log('publicAPIs.manageSrc videoSources.large= ', videoSources.large);
         console.log('publicAPIs.manageSrc videoSources[breakpoint]= ', videoSources[breakpoint]);
+
+        // if activeSource URL is the same URL for this breakpoint, do nothing
         if ( activeSource !== videoSources[breakpoint][0]) {
             console.info('activeSource !== videoSources[breakpoint][1].  activeSource');
             console.info('videoSources[breakpoint][0] = ', videoSources[breakpoint][0]);
             console.info('videoSources[breakpoint][1] 1 = ', videoSources[breakpoint][1]);
+
+            // let's change the src of the source element
             videoSourceElement.setAttribute('src', videoSources[breakpoint][0]);
             //  test if is possible avoid a new load by check if was loaded before
+
+            // loads videos
             videoElement.load();
+
+            // let's signal this video is loaded
+            // TODO: this should be done only after the video fully loads
             videoSources[breakpoint][1] = 'loaded'
+
+            // play it, monkey!
             videoElement.play();
+
+            // update activeSource
             activeSource = videoSources[breakpoint][0];
             console.info('videoSources[breakpoint][1] 2 = ', videoSources[breakpoint][1]);
 
             console.info('manageSrc activeSource after = ', activeSource);
+            // activeVideoBreakpoint
             activeVideoBreakpoint = breakpoint;
         } else {
-            console.info('activeSource === videoSources[breakpoint][1].  activeSource = ');
+            console.info('we don\'t need to change src attribute');
         }
 
     }
 
-    publicAPIs.changeSrc = function(src){
-
-    }
-
+    // init method
     publicAPIs.init = function (options) {
         console.log('publicAPIs.init init ');
+
+        // defaults
+        // TODO merge user options
+        // TODO allows overriding of breakpoint weights
         var defaults = {
             serviceWorker: true
         };
 
+        // if the Service Worker is supported, install it to better cache management
         if (true === defaults.serviceWorker && 'serviceWorker' in navigator) {
             //installWorker();
             console.log('installWorker');
         }
 
+        // run the manageSrc once on load
         publicAPIs.manageSrc(breakpoint);
 
         // Listen for resize events
+        // see https://gomakethings.com/debouncing-events-with-requestanimationframe-for-better-performance/
         window.addEventListener('resize', function (event) {
-
-            console.log('no debounce');
 
             // If there's a timer, cancel it
             if (timeout) {
@@ -237,8 +254,7 @@ var videoElementSrcset = function (videoElement, videoSources) {
             // Setup the new requestAnimationFrame()
             timeout = window.requestAnimationFrame(function () {
 
-                // Run our scroll functions
-                console.log('debounced');
+                // Run our scroll listener
                 breakpointListener();
 
             });
@@ -247,84 +263,6 @@ var videoElementSrcset = function (videoElement, videoSources) {
 
     }
 
-
-    publicAPIs.mediaCheck = function (media) {
-
-        console.info('publicAPIs.mediaCheck init');
-        console.info('publicAPIs.mediaCheck init media = ', media);
-        console.info('publicAPIs.mediaCheck init activeSource = ', activeSource);
-        console.info('publicAPIs.mediaCheck init videoSources[media][1] = ', videoSources[media][1]);
-
-        if (!media || '' === media) {
-            console.error('breakpoint error');
-            return;
-        }
-
-        if (window.matchMedia(media).matches) {
-            console.info('media matches = ', media);
-            if (activeSource !== videoSources[media][0]) {
-                console.info('activeSource !== videoSources[media][1].  activeSource');
-                console.info('videoSources[media][0] = ', videoSources[media][0]);
-                console.info('videoSources[media][2] 1 = ', videoSources[media][2]);
-                videoSourceElement.setAttribute('src', videoSources[media][0]);
-                //  test if is possible avoid a new load by check if was loaded before
-                videoElement.load();
-                videoElement.play();
-                activeSource = videoSources[media][0];
-                console.info('videoSources[media][2] 2 = ', videoSources[media][2]);
-
-                console.info('publicAPIs.mediaCheck activeSource after = ', activeSource);
-            } else {
-                console.info('activeSource === videoSources[media][1].  activeSource = ');
-            }
-
-        } else {
-            console.info('media does not match = ', media);
-        }
-
-    }
-
-    publicAPIs.init2 = function (options) {
-        console.log('publicAPIs.init init ');
-        var defaults = {
-            serviceWorker: true
-        };
-
-        if ( true === defaults.serviceWorker && 'serviceWorker' in navigator ){
-            //installWorker();
-            console.log('installWorker');
-        }
-
-        var mediaQueriesList = [];
-
-        //var mediaQueries = window.matchMedia(mediaQueryList[1]); // Create the query list.
-
-        function handleOrientationChange(evt) {
-            console.log('breakpoint mediaQueryList init for evt = ', evt);
-            console.log('evt.media no init = ', evt.media);
-            if (evt.matches) {
-                console.log('breakpoint matches ');
-                publicAPIs.mediaCheck(evt.media);
-            } else {
-                console.log('breakpoint does not matche');
-                publicAPIs.mediaCheck(evt.media);
-            }
-
-        }// Define a callback function for the event listener.
-
-        // loops through our mediaQueries list and add matchMedia
-        for (i = 0; i < mediaQueries.length; ++i) {
-            mediaQueriesList.push(window.matchMedia(mediaQueries[i]));
-        }
-
-        // loops through our mediaQueriesList, runs handleOrientationChange once and adds
-        // event listener
-        for (i = 0; i < mediaQueriesList.length; ++i) {
-            handleOrientationChange(mediaQueriesList[i]);
-            mediaQueriesList[i].addListener(handleOrientationChange) // call handler function whenever the media query is triggered
-        }
-        console.log('publicAPIs.init end ');
-    }
-
+    // return ou publis APIs
     return publicAPIs;
 }();